+++ packages/frontend/src/components/ThemeToggle.tsx
@@ -1,6 +1,5 @@
-import React from 'react'
 import { Sun, Moon, Monitor } from 'lucide-react'
-import { useTheme } from '../contexts/ThemeContext'
+import { useTheme } from '../hooks/useTheme'
 
 export function ThemeToggle() {
     const { theme, setTheme } = useTheme()
@@ -19,10 +18,9 @@ export function ThemeToggle() {
                     onClick={() => setTheme(value)}
                     className={`
                         flex items-center gap-1 px-2 py-1 rounded-md text-sm font-medium transition-colors
-                        ${
-                            theme === value
-                                ? 'bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 shadow-sm'
-                                : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-100'
+                        ${theme === value
+                            ? 'bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 shadow-sm'
+                            : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-100'
                         }
                     `}
                     aria-label={`Switch to ${label.toLowerCase()} theme`}
