+++ packages/frontend/src/contexts/theme.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react'
+
+export type Theme = 'light' | 'dark' | 'system'
+
+export interface ThemeContextType {
+    theme: Theme
+    effectiveTheme: 'light' | 'dark'
+    setTheme: (theme: Theme) => void
+    toggleTheme: () => void
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined)
